+++ mutwo/music_parameters/__init__.py
@@ -4,10 +4,11 @@ from . import abc
 
 from .ambituses import *
 from .commas import *
+from .lyrics import *
 from .pitches import *
 from .volumes import *
 from .notation_indicators import *
 from .playing_indicators import *
 
 # Force flat structure
-del ambituses, commas, pitches, volumes, notation_indicators, playing_indicators
+del ambituses, commas, lyrics, pitches, volumes, notation_indicators, playing_indicators